Refactor utils.py into three mode-specific modules
Split the monolithic utils.py (2309 lines, 49 functions) into three
focused modules based on functionality:

1. utils_common.py (29 functions)
   - Core utilities used across all modes
   - Version detection, file operations, locale handling
   - ROCm library resolution, UUID generation
   - Getter/setter pattern for rocprof_cmd global

2. utils_profile.py (16 functions)
   - Profile mode specific functionality
   - ROCProf v3 JSON/CSV processing
   - Profiler execution (run_prof, pc_sampling_prof)
   - Kokkos trace processing, sysinfo generation

3. utils_analysis.py (11 functions + 2 classes)
   - Analysis mode specific functionality
   - Workload validation, counter imputation
   - Data aggregation and kernel statistics
   - CallTreeNode and KernelStats dataclasses

Key changes:
- Replaced global rocprof_cmd/rocprof_args with getter/setter pattern
- Updated all imports across 49 Python files
- Fixed all test mocking to target correct modules
- Added proper error handling in rocprof-compute entry point
- Updated copyright year to 2026
- All 287 tests passing

Import audit confirmed:
- 212 utils import statements across codebase
- 0 problematic imports from old utils.utils
- All imports correctly reference refactored modules
