+++ projects/rocprofiler-compute/src/rocprof-compute
@@ -25,14 +25,14 @@
 # SOFTWARE.
 ##############################################################################el
 
+import re
+import sys
+import traceback
 import warnings
+from pathlib import Path
 
 # Suppress pandas FutureWarning about chained assignment
 warnings.filterwarnings("ignore", category=FutureWarning)
-
-import re
-import sys
-from pathlib import Path
 
 try:
     from importlib import metadata
@@ -40,7 +40,7 @@ try:
     from rocprof_compute_base import RocProfCompute
     from utils.logger import console_error
 except ImportError as e:
-    print (f"{e}")
+    print(f"{e}")
     # In wheel package, softlink is not supported.
     # rocprof-compute will get installed in bin and libexec/rocprofiler-compute.
     # When invoked from bin folder, the extra paths are required to import the
@@ -54,12 +54,19 @@ except ImportError as e:
         from importlib import metadata
 
         from rocprof_compute_base import RocProfCompute
-        from utils.utils import console_error
+        from utils.logger import console_error
     except ImportError:
-        pass
+        print(
+            "\n[ERROR] Could not import rocprofiler-compute modules. "
+            "Please check your installation."
+        )
+        script_path = Path(__file__).resolve().parent
+        print(f"Searched: {script_path}, {additional_path.resolve()}")
+        traceback.print_exc()
+        sys.exit(1)
 
 
-def check_version(local_ver, desired_ver, operator) -> bool:
+def check_version(local_ver: str, desired_ver: str, operator: str) -> bool:
     """Check package version strings with simple operators used in companion
     requirements.txt file"""
     ops = {
@@ -79,7 +86,7 @@ def verify_deps() -> None:
     dependencies are not available."""
 
     # Check which version of python is being used
-    if sys.version_info < (3, 8):
+    if sys.version_info < (3, 8):  # noqa: UP036
         print(
             f"[ERROR] Python 3.8 or higher is required to run rocprofiler-compute. "
             f"The current version is {sys.version_info[0]}.{sys.version_info[1]}."
