+++ projects/rocprofiler-compute/src/rocprof_compute_analyze/analysis_cli.py
@@ -35,7 +35,7 @@ from utils import file_io, parser, schema, tty
 from utils.kernel_name_shortener import kernel_name_shortener
 from utils.logger import console_error, console_log, console_warning, demarcate
 from utils.roofline_calc import calc_ai_analyze, validate_roofline_csv
-from utils.utils import (
+from utils.utils_analysis import (
     build_call_trees,
     build_call_trees_with_kernel_ids,
     process_torch_trace_output,
