+++ projects/rocprofiler-compute/src/rocprof_compute_tui/tui_app.py
@@ -40,7 +40,7 @@ import config
 from rocprof_compute_tui.config import APP_TITLE
 from rocprof_compute_tui.views.main_view import MainView
 from utils.specs import generate_machine_specs
-from utils.utils import get_version
+from utils.utils_common import get_version
 
 
 class RocprofTUIApp(App):
